- Expose jwt public key via rest api for clients to consume dynamically - Bring standalone and cluster webservers into closer alignment - Wire up GormProvisioner and RaftProvisioner - Add /farms endpoint - Add /provisioner endpoints - Heavily refactor builders, farm and device factories - Add provisioning / deprovisioning endpoints - Enhancements to JsonWebTokenService / orgs / users / roles - Add JWT refresh token endpoint - Clean up compiler warnings - Add locks to raft state machines (will be removed in next checkin) - Add JWT refresh token endpoint - Get standalone and google auth mostly working with android app

+++ config/channel.go
@@ -1,7 +1,7 @@
 package config
 
 type Channel struct {
-	ID            int         `gorm:"primary_key;AUTO_INCREMENT" yaml:"id" json:"id"`
+	ID            uint64      `gorm:"primary_key;AUTO_INCREMENT" yaml:"id" json:"id"`
 	DeviceID      uint64      `yaml:"device" json:"device_id"`
 	ChannelID     int         `yaml:"channel" json:"channel_id"`
 	Name          string      `yaml:"name" json:"name"`
@@ -22,11 +22,11 @@ func NewChannel() *Channel {
 		Schedule:   make([]Schedule, 0)}
 }
 
-func (channel *Channel) SetID(id int) {
+func (channel *Channel) SetID(id uint64) {
 	channel.ID = id
 }
 
-func (channel *Channel) GetID() int {
+func (channel *Channel) GetID() uint64 {
 	return channel.ID
 }
 
@@ -38,10 +38,12 @@ func (channel *Channel) SetDeviceID(id uint64) {
 	channel.DeviceID = id
 }
 
+// Sets the board / channel id using the ID printed on the PCB
 func (channel *Channel) SetChannelID(id int) {
 	channel.ChannelID = id
 }
 
+// Sets the board / channel id using the ID printed on the PCB
 func (channel *Channel) GetChannelID() int {
 	return channel.ChannelID
 }
