+++ config/schedule.go
@@ -9,7 +9,7 @@ import (
 type Schedule struct {
 	ID             uint64     `gorm:"primary_key;AUTO_INCREMENT" yaml:"id" json:"id"`
 	WorkflowID     uint64     `yaml:"workflow" json:"workflow_id"`
-	ChannelID      int        `yaml:"channelId" json:"channel_id"`
+	ChannelID      uint64     `yaml:"channelId" json:"channel_id"`
 	StartDate      time.Time  `yaml:"startDate" json:"startDate"`
 	EndDate        *time.Time `yaml:"endDate" json:"endDate"`
 	Frequency      int        `yaml:"frequency" json:"frequency"`
@@ -41,11 +41,11 @@ func (schedule *Schedule) GetWorkflowID() uint64 {
 	return schedule.WorkflowID
 }
 
-func (schedule *Schedule) SetChannelID(id int) {
+func (schedule *Schedule) SetChannelID(id uint64) {
 	schedule.ChannelID = id
 }
 
-func (schedule *Schedule) GetChannelID() int {
+func (schedule *Schedule) GetChannelID() uint64 {
 	return schedule.ChannelID
 }
