+++ state/generic_store.go
@@ -5,14 +5,14 @@ import (
 )
 
 type GenericStateStore struct {
-	items map[int]interface{}
+	items map[uint64]interface{}
 	mutex *sync.RWMutex
 	StateStore
 }
 
 func NewGenericStore(len int) StateStore {
 	return &GenericStateStore{
-		items: make(map[int]interface{}, len),
+		items: make(map[uint64]interface{}, len),
 		mutex: &sync.RWMutex{}}
 }
 
@@ -20,13 +20,13 @@ func (store *GenericStateStore) Len() int {
 	return len(store.items)
 }
 
-func (store *GenericStateStore) Put(id int, v interface{}) {
+func (store *GenericStateStore) Put(id uint64, v interface{}) {
 	store.mutex.Lock()
 	defer store.mutex.Unlock()
 	store.items[id] = v
 }
 
-func (store *GenericStateStore) Get(id int) (interface{}, bool) {
+func (store *GenericStateStore) Get(id uint64) (interface{}, bool) {
 	store.mutex.RLock()
 	defer store.mutex.RUnlock()
 	if item, ok := store.items[id]; ok {
